requirements.txt
pytest>=2.2
PyYAML>=3.11
